basic right joystick control for pan and tilt servos
Added support for controlling two servos (pan and tilt) using the right joystick.
Deadzone filtering is applied to avoid jittering when the joystick is idle.
Mapped X-axis to horizontal (pan) movement and Y-axis to vertical (tilt) movement.
Pressing L1 centers both servos to 90 degrees. Servos are properly attached during setup.

But there is a problem my y-axis i broken and gives weird values even when its not touched so I might have to change to using dpads for the vertical part of the servo, it currently works alright but can sometime move by itself cause of the dammage. But horizontal part works really good for now.